requirements.txt
praw
streamlit
plotly
pandas
python-dotenv
nltk
requests
